Cloud/firestoretrigger/requirements.txt
# Function dependencies, for example:
# package>=version
Flask==2.0.1
jsonify==0.5

numpy==1.19.5
tensorflow==2.5.0
google-cloud-storage==1.16.1
h5py==2.10.0

google-cloud-firestore==2.1.1